# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Pedidos de Restaurante
  version: 1.0.0
  description: API para gestionar pedidos, platos, mesas y usuarios de un restaurante.
servers:
  - url: http://localhost:3000/api
    description: Servidor de Desarrollo Local
...
paths:
...
  /orders:
    get:
      summary: Obtener todos los pedidos
      tags: [Pedidos]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pedidos obtenida exitosamente
        '403':
          description: Acceso denegado
    post:
      summary: Crear un nuevo pedido (mesero, admin)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_mesa
                - detalles
              properties:
                id_mesa:
                  type: integer
                  example: 1
                detalles:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderDetailInput'
                estado:
                  type: string
                  enum: [borrador, pendiente, en preparación, servido]
                  example: pendiente
      responses:
        '201':
          description: Pedido creado exitosamente
        '400':
          description: Datos inválidos
        '403':
          description: Acceso denegado
  /orders/{id}:
    get:
      summary: Obtener un pedido por ID
      tags: [Pedidos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del pedido
      responses:
        '200':
          description: Detalles del pedido
        '404':
          description: Pedido no encontrado
    put:
      summary: Actualizar un pedido (mesero, admin)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                detalles:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderDetailInput'
                estado:
                  type: string
                  enum: [borrador, pendiente, en preparación, servido]
                  example: en preparación
      responses:
        '200':
          description: Pedido actualizado exitosamente
        '400':
          description: Datos inválidos
        '403':
          description: Acceso denegado
        '404':
          description: Pedido no encontrado
    delete:
      summary: Eliminar un pedido (solo admin)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del pedido
      responses:
        '200':
          description: Pedido eliminado exitosamente
        '403':
          description: Acceso denegado
        '404':
          description: Pedido no encontrado
  /orders/{id}/status:
    put:
      summary: Actualizar el estado de un pedido (mesero, admin, cocinero)
      tags: [Pedidos]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - estado
              properties:
                estado:
                  type: string
                  enum: [borrador, pendiente, en preparación, servido]
                  example: servido
      responses:
        '200':
          description: Estado del pedido actualizado exitosamente
        '400':
          description: Datos inválidos
        '403':
          description: Acceso denegado
        '404':
          description: Pedido no encontrado
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderDetailInput:
      type: object
      required:
        - id_plato
        - cantidad
      properties:
        id_plato:
          type: integer
          example: 1
        cantidad:
          type: integer
          example: 2
security:
  - bearerAuth: []